Allow logging with Tor browser and strict settings
Under certain strict settings Tor browser does not like it when you use
sessionStorage and/or localStorage.

This diff removes any reference to such storages and uses instead a
dictionary to keep authentication information.

With this diff committed on my staging system I was able to register
myself for ADINA'15 from Tor browser.

This was joint work with @alemela and @davidellavena.