Allow new fields without touching heroku backend
This is done by storing the new fields inside the twitter field
that now becomes the serialization of a JSON.

The code does not forget about a previously existing twitter field
but rather keeps it when you login again.

Tested on the staging system. Should allow us to implement the
new functionality without updating the backend on heroku.